Fix use of deprecated parser behavior
gpxpy/parser.py:228 uses a test for boolean falseness to check if an xml
node is not found:

"""
  if not node:
"""

With python-lxml 2.3.5, this produces a warning:
./gpxpy/gpxpy/parser.py:228: FutureWarning: The behavior of this method
will change in future versions. Use specific 'len(elem)' or 'elem is not
None' test instead.

This patch just changes the test to:
"""
  if node is None:
"""

Unit tests pass.