Ignore empty track segment when calculating bounds
In situations where an empty track segment occurs at the end of a
track file the bounding box calculation would produce None values
for the minimum latitude and longitude because None < x evaluates
to true. By explicitely checking if the track segment has a valid
bounding box this can be circumvented easily.