Don't lose the second decimals from the timestamps
The time parsing code was losing the second decimals. When the GPX file
is used to interpolate the position of photos based on their timestamp
this can lead to a loss of precision proportional to the speed at which
the camera is moving, that is as much as 30 meters at highway speeds.
Only drop the '+HH:MM' and 'Z' timezone specifications when they are at
the end of the string and truncate the second decimals to microsecond
precision to accomodate datetime's strptime() function.
Modify test_parse_time() so it verifies the returned datetime object,
including the second decimals unlike test_long_timestamps()!
Modify TimeConverter.to_string() so it preserves the sub-second data but
preserves the old format if none is specified.

Signed-off-by: Francois Gouget <fgouget@free.fr>