Beef up the time parsing tests
Test the timestamp parsing function through TimeConverter class which is
the actual interface used by gpxpy. This allows testing the conversion
of DateTime objects back to strings.
Remove the code that's stripping the trailing 'Z' from the test
strings: it's shorter and clearer to just add the relevant tests.
For each timestamp store the actual DateTime components we're expecting
so we can more precisely check the result of the conversion.
Add more variants of the timestamp formats we want to support.

Signed-off-by: Francois Gouget <fgouget@free.fr>