Return the timezone information if present
Applications that use GPX files to geolocalize other data (e.g. photos)
based on the timestamps must know which timezone each of them is basedi
on. So return the timezone information if specified in the timestamp.

Note that this changes the gpxpy API in that timezone aware DateTime
objects cannot be directly compared with naive DateTime objects. This
means applications that only dealt with naive DateTime objects will now
need to at least be ready to strip the tzinfo field of DateTime objects
returned by gpxpy.

Signed-off-by: Francois Gouget <fgouget@free.fr>